+++ CHANGELOG.md
@@ -1,8 +1,14 @@
+## [Unreleased]
+
+### Added
+
+- Implemented `Display` for `SyncOutput`.
+
 ## [0.1.3] - 2026-06-19
 
 ### Added
 
-- Automatic changelog version updated (is this thing on?)
+- Automatic changelog version updated (is this thing on?).
 
 ## [0.1.2] - 2026-19-06